requirements.txt
google-cloud-storage==2.18.2
spacy==3.7.5
functions-framework==3.8.1
